refactor: renomeia a classe Cadastro para Cadastrar e move seu conteúdo para um novo arquivo
